core:util 모듈 추가
ui 모듈과 util 모듈을 분리, 다른 모듈들에서 공통으로 사용하는 ui와 관련되지 않는 함수 및 클래스를 모아둠
패키지 구조 변경-> 확장함수가 아닌 파일들은 extension 밖으로 꺼냄
