+++ core/ui/src/main/kotlin/com/nexters/bandalart/android/core/ui/component/EmojiText.kt
@@ -3,7 +3,7 @@ package com.nexters.bandalart.android.core.ui.component
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.TextUnit
-import com.nexters.bandalart.android.core.ui.extension.nonScaleSp
+import com.nexters.bandalart.android.core.ui.nonScaleSp
 
 @Composable
 fun EmojiText(
