+++ core/util/src/main/kotlin/com/nexters/bandalart/android/core/util/extension/Context.kt
@@ -1,0 +1,8 @@
+package com.nexters.bandalart.android.core.util.extension
+
+import android.content.Context
+import java.util.Locale
+
+fun Context.getCurrentLocale(): Locale {
+  return this.resources.configuration.locales.get(0)
+}
