+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/HomeViewModel.kt
@@ -16,7 +16,7 @@ import com.nexters.bandalart.android.core.domain.usecase.bandalart.ShareBandalar
 import com.nexters.bandalart.android.core.domain.usecase.bandalart.UpdateBandalartEmojiUseCase
 import com.nexters.bandalart.android.core.domain.usecase.bandalart.UpsertBandalartKeyUseCase
 import com.nexters.bandalart.android.core.ui.R
-import com.nexters.bandalart.android.core.ui.extension.UiText
+import com.nexters.bandalart.android.core.ui.UiText
 import com.nexters.bandalart.android.feature.home.mapper.toEntity
 import com.nexters.bandalart.android.feature.home.mapper.toUiModel
 import com.nexters.bandalart.android.feature.home.model.BandalartCellUiModel
