+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/ui/HomeTopBar.kt
@@ -27,7 +27,7 @@ import com.nexters.bandalart.android.core.designsystem.theme.White
 import com.nexters.bandalart.android.core.designsystem.theme.pretendard
 import com.nexters.bandalart.android.core.ui.R
 import com.nexters.bandalart.android.core.ui.component.FixedSizeText
-import com.nexters.bandalart.android.core.ui.extension.nonScaleSp
+import com.nexters.bandalart.android.core.ui.nonScaleSp
 
 @Composable
 internal fun HomeTopBar(
