+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/ui/bandalart/BandalartCellGrid.kt
@@ -8,7 +8,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.nexters.bandalart.android.core.ui.extension.ThemeColor
+import com.nexters.bandalart.android.core.ui.ThemeColor
 
 @Composable
 fun BandalartCellGrid(
