+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/ui/bandalart/BandalartColorPicker.kt
@@ -18,12 +18,12 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.nexters.bandalart.android.core.ui.extension.ThemeColor
-import com.nexters.bandalart.android.core.ui.extension.allColor
 import com.nexters.bandalart.android.core.ui.extension.noRippleClickable
-import com.nexters.bandalart.android.core.ui.extension.toColor
 import com.nexters.bandalart.android.core.designsystem.theme.Gray900
 import com.nexters.bandalart.android.core.designsystem.theme.White
+import com.nexters.bandalart.android.core.ui.ThemeColor
+import com.nexters.bandalart.android.core.ui.allColor
+import com.nexters.bandalart.android.core.ui.extension.toColor
 
 @Composable
 fun BandalartColorPicker(
