+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/ui/bandalart/BandalartEmojiPicker.kt
@@ -34,7 +34,7 @@ import com.nexters.bandalart.android.core.designsystem.theme.Gray100
 import com.nexters.bandalart.android.core.designsystem.theme.Gray400
 import com.nexters.bandalart.android.core.designsystem.theme.White
 import com.nexters.bandalart.android.core.ui.component.EmojiText
-import com.nexters.bandalart.android.core.ui.extension.NavigationBarHeightDp
+import com.nexters.bandalart.android.core.ui.NavigationBarHeightDp
 import java.util.*
 import kotlinx.coroutines.launch
